Working App, in progress
The app is now able get weather data using the fetch api. I had trouble getting the API key to work but figured it out. The code is refactored now in a client folder and a server folder. Will continue studying the code tomorrow and updating the UI. Then moving on to testing.